fix: apply roborev fix for 9b77516 (job #80)
