+++ src/lib.rs
@@ -1,3 +1,5 @@
+use actix_web_prom::{PrometheusMetrics, PrometheusMetricsBuilder};
+
 pub mod db;
 pub mod errors;
 pub mod handlers;
@@ -6,3 +8,13 @@ pub mod openapi;
 pub mod routes;
 pub mod schema;
 pub mod serializers;
+
+pub fn build_prometheus() -> PrometheusMetrics {
+    PrometheusMetricsBuilder::new("api")
+        .endpoint("/metrics")
+        .exclude("/metrics")
+        .exclude("/health")
+        .mask_unmatched_patterns("UNKNOWN")
+        .build()
+        .expect("Failed to initialize Prometheus metrics")
+}
